debug

+++ video/dataset.py
@@ -5,7 +5,7 @@ import numpy as np
 
 class VideoMnistDataset(Dataset):
     def __init__(self, path, frame_len, start_sample, end_sample):
-        self.frame_len = frame_len
+        self.frame_len = int(frame_len)
         self.frames = np.load(path)
         self.frames = self.frames.swapaxes(0, 1)[:, :, start_sample:end_sample, :]
         frames_shape = self.frames.shape
@@ -18,6 +18,7 @@ class VideoMnistDataset(Dataset):
         return self.length
 
     def __getitem__(self, index):
-        video_ind = index / self.sample_per_video
+        video_ind = int(index / self.sample_per_video)
         frame_ind = index - video_ind * self.sample_per_video
+
         return self.frames[video_ind, frame_ind: frame_ind + self.frame_len, :, :]
